Revert "Unit test channel"
