+++ .eslintrc.json
@@ -2,8 +2,7 @@
     "env": {
         "es2020": true,
         "node": true,
-        "es6": true,
-        "jest": true
+        "es6": true
     },
     "extends": [
         "airbnb-base",
