+++ src/screens/ChannelListScreen/index.js
@@ -26,11 +26,10 @@ import { useAfterInteractions } from '../../hooks/useAfterInteractions';
 import { withInteractionsManaged } from '../../components/WithInteractionManaged';
 import CustomPreviewUnreadCount from './elements/CustomPreviewUnreadCount';
 import PostNotificationPreview from './elements/components/PostNotificationPreview';
-import { getSpecificCache } from '../../utils/cache';
-import { FEED_COMMENT_COUNT } from '../../utils/cache/constant';
+import { getSpecificCache, saveToCache } from '../../utils/cache';
+import { CHAT_FOLLOWING_COUNT, FEED_COMMENT_COUNT } from '../../utils/cache/constant';
 import PreviewMessage from './elements/CustomPreviewMessage';
 import { setTotalUnreadPostNotif } from '../../context/actions/unReadMessageAction';
-import useChannelList from './hooks/useChannelList';
 
 
 const ChannelListScreen = ({ navigation }) => {
@@ -48,9 +47,9 @@ const ChannelListScreen = ({ navigation }) => {
   const [countReadComment, setCountReadComment] = React.useState({})
   // const [countChat, setCountChat] = React.useState({})
   const {myProfile} = profileContext
-  const {mappingUnreadCountPostNotifHook, handleNotHaveCacheHook, handleUpdateCacheHook} = useChannelList()
+  const [postCount, setPostCount] = React.useState(0)
 
-  const [, dispatchUnreadMessage] =
+  const [unReadMessage, dispatchUnreadMessage] =
     React.useContext(Context).unReadMessage;
 
   const filters = {
@@ -119,19 +118,36 @@ const handleCacheComment  = () => {
     }
   })
 }
-const handleNotHaveCache = async () => {
-  const comment =  await handleNotHaveCacheHook(listPostNotif)
+const handleNotHaveCache = () => {
+  let comment = {}
+  listPostNotif.forEach((data) => {
+    comment = {...comment, [data.activity_id]: 0}
+  })
+  saveToCache(FEED_COMMENT_COUNT, comment)
   setCountReadComment(comment)
 }
 
-const handleUpdateCache = async (id, totalComment) => {
-  const updateReadCache = await handleUpdateCacheHook(countReadComment, id, totalComment)
+const handleUpdateCache = (id, totalComment) => {
+  const updateReadCache = {...countReadComment, [id]: totalComment}
+  saveToCache(FEED_COMMENT_COUNT, updateReadCache)
   setCountReadComment(updateReadCache)
 }
-const mappingUnreadCountPostNotif = async () => {
-  const totalMessage = await mappingUnreadCountPostNotifHook(listPostNotif, countReadComment)
-  dispatchUnreadMessage(setTotalUnreadPostNotif(totalMessage))
+console.log(countReadComment, 'papana')
+const mappingUnreadCountPostNotif = (readComment = countReadComment) => {
+  if(listPostNotif.length > 0) {
+      const maping = listPostNotif.map((notif) => ({id: notif.activity_id, totalComment: notif.totalCommentBadge}))
+      const mapCount = maping.map((data) => ({
+            ...data,
+            totalComment: data.totalComment -( countReadComment[data.id] || 0)
+          }))
+      const countTotal = mapCount.map((count) => count.totalComment).reduce((a, b) => a + b)
+      const totalMessage = countTotal
+      dispatchUnreadMessage(setTotalUnreadPostNotif(totalMessage))
+  }
+
 }
+
+console.log(postCount, 'mamang')
 
   const getPostNotification = async () => {
     const res = await getFeedNotification()
